requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via uvicorn
dnspython==2.3.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.0.0.post2
    # via fastapi
fastapi[all]==0.95.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.0
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.0
    # via fastapi
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
loguru==0.7.0
    # via -r requirements.in
markupsafe==2.1.2
    # via jinja2
orjson==3.8.11
    # via fastapi
pydantic==1.10.7
    # via fastapi
pymongo[srv]==4.3.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.6
    # via fastapi
pyyaml==6.0
    # via
    #   fastapi
    #   uvicorn
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via pydantic
ujson==5.7.0
    # via fastapi
uvicorn[standard]==0.22.0
    # via fastapi
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.2
    # via uvicorn
